divided the views for different controllers
