+++ Messanger_Test/AppDelegate.swift
@@ -23,8 +23,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.rootViewController = navigationController
         window?.makeKeyAndVisible()
         
-
-        
         return true
     }
